* feature/FlightHistory:
  [Update] Show the drone location TextView in PLANNING & FOLLOW ME
  [Remove] UI remove Follow Me
  [Fix] Avoid Null Exception
  [Remove] Mission List 1st mission point

+++ app/src/main/java/com/coretronic/drone/missionplan/fragments/MavInfoFragment.java
@@ -12,6 +12,6 @@ public class MavInfoFragment extends Fragment{
     public void setMavInfoSpeed(float groundSpeed) {
     }
 
-    public void setMavInfoLocation(double droneLat, double droneLng) {
+    public void setMavInfoLocation(long droneLat, long droneLng) {
     }
 }
